# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My swager Doc",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get prepared info about order",
                "operationId": "5f18eef3bb3eb0fdb460a5a9ad6348f4",
                "responses": {
                    "200": {
                        "description": "prepare order",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Jhon"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "jhon@mail.ru"
                                                },
                                                "phone_number": {
                                                    "type": "string",
                                                    "example": "89617756027"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "locations": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "New York"
                                                    },
                                                    "street": {
                                                        "type": "string",
                                                        "example": "Lenina"
                                                    },
                                                    "house_number": {
                                                        "type": "string",
                                                        "example": "25b"
                                                    },
                                                    "floor": {
                                                        "type": "integer",
                                                        "example": 15
                                                    },
                                                    "apartment": {
                                                        "type": "integer",
                                                        "example": 151
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "cart": {
                                            "properties": {
                                                "products": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "product_name": {
                                                                "type": "string",
                                                                "example": "Something tasty"
                                                            },
                                                            "size": {
                                                                "type": "string",
                                                                "example": "Some size"
                                                            },
                                                            "price": {
                                                                "type": "integer",
                                                                "example": 599
                                                            },
                                                            "quantity": {
                                                                "type": "integer",
                                                                "example": 1
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "total_price": {
                                                    "type": "integer",
                                                    "example": 1000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "example": "Ooops! Looks like something went wrong."
                                        },
                                        "message": {
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create order",
                "operationId": "c25ee7942d4fda0a787e13dd69e5bb90",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "required": [
                                            "phone_number",
                                            "location_id"
                                        ],
                                        "properties": {
                                            "phone_number": {
                                                "type": "string",
                                                "example": "89617756027"
                                            },
                                            "location_id": {
                                                "type": "integer",
                                                "example": 2
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Order created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Empty cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "example": "Ooops! Looks like something went wrong."
                                        },
                                        "message": {
                                            "example": "Your cart is empty.."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/own-orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get all own orders",
                "operationId": "96f5e3cb52f9b5fb835acdc2039f5364",
                "responses": {
                    "200": {
                        "description": "prepare order",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "orders": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Jhon"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "jhon@mail.ru"
                                                            },
                                                            "phone_number": {
                                                                "type": "string",
                                                                "example": "89617756027"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2025-01-23 21:20:45"
                                                    },
                                                    "locations": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "city": {
                                                                "type": "string",
                                                                "example": "New York"
                                                            },
                                                            "street": {
                                                                "type": "string",
                                                                "example": "Lenina"
                                                            },
                                                            "house_number": {
                                                                "type": "string",
                                                                "example": "25b"
                                                            },
                                                            "floor": {
                                                                "type": "integer",
                                                                "example": 15
                                                            },
                                                            "apartment": {
                                                                "type": "integer",
                                                                "example": 151
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "products": {
                                                        "properties": {
                                                            "total_price": {
                                                                "type": "integer",
                                                                "example": 1000
                                                            }
                                                        },
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "product_name": {
                                                                    "type": "string",
                                                                    "example": "Something tasty"
                                                                },
                                                                "size": {
                                                                    "type": "string",
                                                                    "example": "Some size"
                                                                },
                                                                "price": {
                                                                    "type": "integer",
                                                                    "example": 599
                                                                },
                                                                "quantity": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "example": "Ooops! Looks like something went wrong."
                                        },
                                        "message": {
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the following format: `Bearer {token}`",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}